tests/wpt/tests/html/dom/render-blocking/element-render-blocking-004.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Removing link in the head has an effect</title>

<link id=link rel=expect href="#last" blocking="render">
<script>
link.remove();

async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "removing link in the head makes it no longer blocking");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
